219. Contains Duplicate II
